pyproject.toml
[tool.poetry]
name = "drf-api-deploy"
version = "0.1.0"
description = ""
authors = ["MohamadSheikhAlshabab <alshabab.moh@gmail.com>"]

[tool.poetry.dependencies]
python = "~3.8"
django = "^3.1.1"
djangorestframework = "^3.12.1"
djangorestframework_simplejwt = "^4.4.0"
psycopg2-binary = "^2.8.6"
environ = "^1.0"
django-environ = "^0.4.5"
whitenoise = "^5.2.0"

[tool.poetry.dev-dependencies]
black = "^20.8b1"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
